phpstan, move exception, crlf->lf
