+++ composer.json
@@ -19,9 +19,9 @@
   "require-dev": {
     "doctrine/coding-standard": "^5.0",
     "phpbench/phpbench": "^0.14.0",
-    "phpstan/phpstan": "0.10.5",
-    "phpstan/phpstan-phpunit": "0.10.0",
-    "phpstan/phpstan-strict-rules": "0.10.1",
+    "phpstan/phpstan": "0.11.16",
+    "phpstan/phpstan-phpunit": "0.11.2",
+    "phpstan/phpstan-strict-rules": "0.11.1",
   	"phpunit/phpcov": "^5.0",
     "phpunit/phpunit": "^7.2",
     "webonyx/graphql-php": "^0.13.0"
