+++ src/Type/Definition/ValidateItemsError.php
@@ -1,0 +1,12 @@
+<?php
+namespace GraphQL\Type\Definition;
+
+class ValidateItemsError extends \Exception {
+	public $path;
+	public $error;
+	function __construct($path, $error) {
+		parent::__construct();
+		$this->path = $path;
+		$this->error = $error;
+	}
+}
